sora_cam_label_detection_with_rekognition/requirements.txt
pillow
requests==2.28.2
soracam-python-client @ git+https://github.com/soracom-labs/sora-cam-python-client@main ; python_version >= "3.9"
